Move types out of BaseSheetViewModel.
