+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/model/PaymentSheetViewState.kt
@@ -1,16 +1,15 @@
 package com.stripe.android.paymentsheet.model
 
 import com.stripe.android.paymentsheet.PaymentSheetActivity
-import com.stripe.android.paymentsheet.viewmodels.BaseSheetViewModel
 
 /**
  * This will show the state of the [PaymentSheetActivity] as it does work.  The states always
  * progress as follows: Ready -> StartProcessing -> FinishProcessing -> ProcessResult
  */
 internal sealed class PaymentSheetViewState(
-    val errorMessage: BaseSheetViewModel.UserErrorMessage? = null
+    val errorMessage: UserErrorMessage? = null
 ) {
-    data class Reset(private val message: BaseSheetViewModel.UserErrorMessage? = null) :
+    data class Reset(private val message: UserErrorMessage? = null) :
         PaymentSheetViewState(message)
 
     data object StartProcessing : PaymentSheetViewState(null)
@@ -18,4 +17,6 @@ internal sealed class PaymentSheetViewState(
     data class FinishProcessing(
         val onComplete: () -> Unit
     ) : PaymentSheetViewState()
+
+    data class UserErrorMessage(val message: String)
 }
