+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/viewmodels/BaseSheetViewModel.kt
@@ -15,12 +15,11 @@ import com.stripe.android.lpmfoundations.paymentmethod.PaymentMethodMetadata
 import com.stripe.android.model.CardBrand
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.PaymentMethodCode
-import com.stripe.android.model.PaymentMethodCreateParams
-import com.stripe.android.model.PaymentMethodExtraParams
 import com.stripe.android.model.PaymentMethodUpdateParams
 import com.stripe.android.model.SetupIntent
 import com.stripe.android.payments.paymentlauncher.PaymentResult
 import com.stripe.android.paymentsheet.LinkHandler
+import com.stripe.android.paymentsheet.NewOrExternalPaymentSelection
 import com.stripe.android.paymentsheet.PaymentOptionsItem
 import com.stripe.android.paymentsheet.PaymentOptionsViewModel
 import com.stripe.android.paymentsheet.PaymentSheet
@@ -678,53 +677,6 @@ internal abstract class BaseSheetViewModel(
     abstract fun onError(@StringRes error: Int? = null)
 
     abstract fun onError(error: String? = null)
-
-    data class UserErrorMessage(val message: String)
-
-    internal sealed interface NewOrExternalPaymentSelection {
-
-        val paymentSelection: PaymentSelection
-
-        fun getPaymentMethodCode(): PaymentMethodCode
-
-        fun getType(): String
-
-        fun getPaymentMethodCreateParams(): PaymentMethodCreateParams?
-
-        fun getPaymentMethodExtraParams(): PaymentMethodExtraParams?
-
-        data class New(override val paymentSelection: PaymentSelection.New) : NewOrExternalPaymentSelection {
-
-            override fun getPaymentMethodCode(): PaymentMethodCode {
-                return when (paymentSelection) {
-                    is PaymentSelection.New.LinkInline -> PaymentMethod.Type.Card.code
-                    is PaymentSelection.New.Card,
-                    is PaymentSelection.New.USBankAccount,
-                    is PaymentSelection.New.GenericPaymentMethod -> paymentSelection.paymentMethodCreateParams.typeCode
-                }
-            }
-
-            override fun getType(): String = paymentSelection.paymentMethodCreateParams.typeCode
-
-            override fun getPaymentMethodCreateParams(): PaymentMethodCreateParams =
-                paymentSelection.paymentMethodCreateParams
-
-            override fun getPaymentMethodExtraParams(): PaymentMethodExtraParams? =
-                paymentSelection.paymentMethodExtraParams
-        }
-
-        data class External(override val paymentSelection: PaymentSelection.ExternalPaymentMethod) :
-            NewOrExternalPaymentSelection {
-
-            override fun getPaymentMethodCode(): PaymentMethodCode = paymentSelection.type
-
-            override fun getType(): String = paymentSelection.type
-
-            override fun getPaymentMethodCreateParams(): PaymentMethodCreateParams? = null
-
-            override fun getPaymentMethodExtraParams(): PaymentMethodExtraParams? = null
-        }
-    }
 
     companion object {
         internal const val SAVED_CUSTOMER = "customer_info"
